etl/requirements.txt
click==8.1.7
delta-spark==3.2.0
ghp-import==2.1.0
importlib_metadata==7.2.1
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
py4j==0.10.9.7
pyspark==3.5.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyyaml_env_tag==0.1
six==1.16.0
watchdog==4.0.1
zipp==3.19.2
